+++ link-move/src/main/java/com/nhl/link/move/runtime/task/create/CreateTask.java
@@ -2,7 +2,7 @@ package com.nhl.link.move.runtime.task.create;
 
 import com.nhl.link.move.CountingRowReader;
 import com.nhl.link.move.Execution;
-import com.nhl.link.move.Row;
+import com.nhl.link.move.RowAttribute;
 import com.nhl.link.move.RowReader;
 import com.nhl.link.move.batch.BatchProcessor;
 import com.nhl.link.move.batch.BatchRunner;
@@ -15,7 +15,6 @@ import com.nhl.link.move.runtime.token.ITokenManager;
 import org.apache.cayenne.DataObject;
 import org.apache.cayenne.ObjectContext;
 
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -56,9 +55,8 @@ public class CreateTask<T extends DataObject> extends BaseTask {
 
         try (Execution execution = new Execution("CreateTask:" + extractorName, params)) {
 
-            BatchProcessor<Row> batchProcessor = createBatchProcessor(execution);
-
             try (RowReader data = getRowReader(execution, params)) {
+                BatchProcessor<Object[]> batchProcessor = createBatchProcessor(execution, data.getHeader());
                 BatchRunner.create(batchProcessor).withBatchSize(batchSize).run(data);
             }
 
@@ -66,16 +64,9 @@ public class CreateTask<T extends DataObject> extends BaseTask {
         }
     }
 
-    protected BatchProcessor<Row> createBatchProcessor(final Execution execution) {
-        return new BatchProcessor<Row>() {
-
-            ObjectContext context = targetCayenneService.newContext();
-
-            @Override
-            public void process(List<Row> rows) {
-                processor.process(execution, new CreateSegment<T>(context, rows));
-            }
-        };
+    protected BatchProcessor<Object[]> createBatchProcessor(Execution execution, RowAttribute[] rowsHeader) {
+        ObjectContext context = targetCayenneService.newContext();
+        return rows -> processor.process(execution, new CreateSegment<T>(context, rowsHeader, rows));
     }
 
     /**