+++ link-move/src/main/java/com/nhl/link/move/runtime/task/sourcekeys/SourceKeysTask.java
@@ -1,13 +1,9 @@
 package com.nhl.link.move.runtime.task.sourcekeys;
 
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-
 import com.nhl.link.move.CountingRowReader;
-import com.nhl.link.move.LmTask;
 import com.nhl.link.move.Execution;
-import com.nhl.link.move.Row;
+import com.nhl.link.move.LmTask;
+import com.nhl.link.move.RowAttribute;
 import com.nhl.link.move.RowReader;
 import com.nhl.link.move.batch.BatchProcessor;
 import com.nhl.link.move.batch.BatchRunner;
@@ -16,6 +12,9 @@ import com.nhl.link.move.extractor.model.ExtractorName;
 import com.nhl.link.move.runtime.extractor.IExtractorService;
 import com.nhl.link.move.runtime.task.BaseTask;
 import com.nhl.link.move.runtime.token.ITokenManager;
+
+import java.util.HashSet;
+import java.util.Map;
 
 /**
  * An {@link LmTask} that extracts all the keys from the source data store. The
@@ -53,9 +52,8 @@ public class SourceKeysTask extends BaseTask {
 
 			execution.setAttribute(RESULT_KEY, new HashSet<>());
 
-			BatchProcessor<Row> batchProcessor = createBatchProcessor(execution);
-
 			try (RowReader data = getRowReader(execution, params)) {
+				BatchProcessor<Object[]> batchProcessor = createBatchProcessor(execution, data.getHeader());
 				BatchRunner.create(batchProcessor).withBatchSize(batchSize).run(data);
 			}
 
@@ -63,14 +61,8 @@ public class SourceKeysTask extends BaseTask {
 		}
 	}
 
-	protected BatchProcessor<Row> createBatchProcessor(final Execution execution) {
-		return new BatchProcessor<Row>() {
-
-			@Override
-			public void process(List<Row> rows) {
-				processor.process(execution, new SourceKeysSegment(rows));
-			}
-		};
+	protected BatchProcessor<Object[]> createBatchProcessor(Execution execution, RowAttribute[] rowsHeader) {
+		return rows -> processor.process(execution, new SourceKeysSegment(rowsHeader, rows));
 	}
 
 	/**