+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #160 Upgrade Jackson to 2.9.5
 * #162 Remove API deprecated <= v.2.4
+* #163 Replace source Row object with just Object[] 
 
 ## Release 2.6